Move ACF fields to OSI plugin
This is an attempt to help with posting Press mentions automatically from a Google sheet via 3rd party automators such as Make or Zapier. At the moment, these services don't recognize the custom fields created by ACF. This change is an attempt to fix or rule out ACF as the cause of the issue.